@@ -12,13 +12,14 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@skeletonlabs/skeleton": "2.9.2",
-    "@skeletonlabs/tw-plugin": "0.4.0",
-    "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
-    "@tailwindcss/forms": "0.5.7",
-    "@tailwindcss/typography": "0.5.12",
+    "@lucide/svelte": "^0.522.0",
+    "@skeletonlabs/skeleton": "^3.1.3",
+    "@skeletonlabs/skeleton-svelte": "^1.2.3",
+    "@sveltejs/adapter-auto": "^6.0.1",
+    "@sveltejs/kit": "^2.22.0",
+    "@sveltejs/vite-plugin-svelte": "^5.1.0",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16",
     "@types/eslint": "8.56.0",
     "@types/node": "^20.12.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
@@ -27,21 +28,20 @@
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.35.1",
-    "postcss": "8.4.38",
     "prettier": "^3.1.1",
     "prettier-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
-    "tailwindcss": "3.4.3",
+    "svelte": "^5.34.7",
+    "svelte-check": "^4.2.2",
+    "tailwindcss": "^4.1.10",
     "tslib": "^2.4.1",
     "typescript": "^5.0.0",
-    "vite": "^5.0.3",
-    "vite-plugin-tailwind-purgecss": "0.3.3"
+    "vite": "^6.3.5"
   },
   "type": "module",
   "dependencies": {
     "@cmi-dair/skeleton-themes": "file:..",
     "@floating-ui/dom": "1.6.3",
+    "@tailwindcss/vite": "^4.1.10",
     "highlight.js": "11.9.0"
   }
 }
