@@ -1,31 +1,12 @@
 <script lang="ts">
-	import { AppBar, LightSwitch, getDrawerStore } from '@skeletonlabs/skeleton';
+	import { AppBar } from '@skeletonlabs/skeleton-svelte';
+	import { Modal } from '@skeletonlabs/skeleton-svelte';
 
-	const drawerStore = getDrawerStore();
-
-	function drawerOpen(): void {
-		drawerStore.open({});
-	}
+	let drawerState = $state(false);
 </script>
 
-<AppBar data-testid="app-bar">
-	<svelte:fragment slot="lead">
-		<div class="flex items-center">
-			<button class="md:hidden btn btn-sm mr-4" on:click={drawerOpen}>
-				<span>
-					<svg viewBox="0 0 100 80" class="fill-token w-4 h-4">
-						<rect width="100" height="20" />
-						<rect y="30" width="100" height="20" />
-						<rect y="60" width="100" height="20" />
-					</svg>
-				</span>
-			</button>
-			<a href="/" data-testid="a-cmi-logo" class="hidden md:block">
-				<img src="/cmi.svg" alt="Agate" class="w-28" />
-			</a>
-		</div>
-	</svelte:fragment>
-	<svelte:fragment slot="trail">
+<AppBar background="bg-surface-50">
+	{#snippet trail()}
 		<button class="btn hover:variant-soft-primary">
 			<a
 				data-testid="a-github"
@@ -33,16 +14,15 @@
 				target="_blank"
 				rel="noreferrer"
 			>
-				<i class="fa-brands fa-github text-lg" />
+				<i class="fa-brands fa-github text-lg"></i>
 				<span>GitHub</span>
 			</a>
 		</button>
 		<button class="btn hover:variant-soft-primary">
 			<a data-testid="a-mailto" href="mailto:dair@childmind.org" target="_blank" rel="noreferrer">
-				<i class="fa-solid fa-envelope text-lg" />
+				<i class="fa-solid fa-envelope text-lg"></i>
 				<span>Contact</span>
 			</a>
 		</button>
-		<LightSwitch />
-	</svelte:fragment>
+	{/snippet}
 </AppBar>
